requirements.txt
dict.sorted>=1.0.0
django>=1.6
djangorestframework>=2.3.10
pytz>=2013.9
six>=1.5.2
